refactor: ViewModel이 Application을 참조하지 않도록 변경
